backend/requirements.txt
Flask
flask-cors
numpy
Pillow
tensorflow
gdown
gunicorn
tf-keras 